+++ examples/lastrev-next-starter/packages/components/src/components/Text/Text.theme.ts
@@ -6,9 +6,13 @@ export const defaultProps: ComponentsProps['Text'] = {};
 // https://mui.com/customization/theme-components/#global-style-overrides
 export const styleOverrides: ComponentsOverrides<Theme>['Text'] = {
   // Set some static styles
-  // root: {
-  //   backgroundColor: 'red'
-  // }
+  root: {
+    'ol, ul, li': {
+      /* Revert padding reset is what gives the indentation to list */
+
+      padding: 'revert'
+    }
+  }
   //
   // Use the ownerState to set dynamic styles
   // root: ({ ownerState, theme }) => {