+++ examples/lastrev-next-starter/packages/web/cypress/support/commands/index.js
@@ -1,5 +1,0 @@
-// import * as generic from './generic';
-
-// Generic
-// Cypress.Commands.add('selectDropdown', generic.selectDropdown);
-import '@percy/cypress';