+++ examples/lastrev-next-starter/packages/web/pages/_document.tsx
@@ -8,10 +8,9 @@ export default class MyDocument extends Document {
     return (
       <Html lang="en">
         <Head>
-          {/* Remove if not needed */}
           <link
-            rel="stylesheet"
             href="https://fonts.googleapis.com/css?family=Open+Sans:300,400,500,700&display=swap"
+            rel="stylesheet"
           />
         </Head>
         <body>