+++ examples/lastrev-next-starter/scripts/dev.sh
@@ -1,12 +1,14 @@
 #!/bin/bash
 
 function cleanup() {
-  rv=$?
-  npx pm2 delete gql-serve
-  exit $rv
+    rv=$?
+    npx pm2 delete gql-serve
+    exit $rv
 }
 
 trap "cleanup" EXIT
 
 echo "Starting develop server..."
-yarn propagate:env && turbo run dev --output-logs=new-only
+yarn propagate:env
+turbo run sync:cms
+turbo run dev --output-logs=new-only