+++ packages/component-library/src/components/Collection/Collection.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { Container, Box } from '@mui/material';
 import styled from '@mui/system/styled';
-import omit from 'lodash/omit';
+
 import ErrorBoundary from '../ErrorBoundary';
 import ContentModule from '../ContentModule';
 import Section from '../Section';
-import sidekick from '../../utils/sidekick';
+import sidekick from '@last-rev/contentful-sidekick-util';
 import ConditionalWrapper from '../ConditionalWrapper';
 import { CollectionProps } from './Collection.types';
 
@@ -29,7 +29,7 @@ export const Collection = ({
         variant={variant}
         itemsVariant={itemsVariant}
         data-testid="Collection"
-        {...omit(props, 'theme')}
+        {...props}
         {...sidekick(sidekickLookup)}
         sx={styles?.root}
       >