+++ packages/component-library/src/components/CollectionAccordionMedia/index.ts
@@ -1,2 +1,2 @@
 export { default } from './CollectionAccordionMedia';
-export type { CollectionAccordionMediaProps } from './CollectionAccordionMedia';
+export type { CollectionAccordionMediaProps } from './CollectionAccordionMedia.types';