+++ packages/component-library/src/components/ContentThemeProvider/index.ts
@@ -1,0 +1,4 @@
+export { default } from './ContentThemeProvider';
+export type { ContentThemeProviderProps } from './ContentThemeProvider.types';
+
+export * from './ContentThemeProvider';