+++ packages/component-library/src/components/Image/Image.tsx
@@ -1,45 +1,61 @@
 import React from 'react';
-import LazyLoad from 'react-lazyload';
 import Head from 'next/head';
-import NextImage from 'next/image';
-import SVG from 'react-inlinesvg';
+// @ts-ignore-next-line
+import NextImage from 'next/future/image';
 import ErrorBoundary from '../ErrorBoundary';
 import getImgSrcTag from '../../utils/getImgSrcTag';
 import { ImageProps } from './Image.types';
+import dynamic from 'next/dynamic';
 
-// TODO: Move this to the correct place
-type Env = 'development' | 'production' | 'test' | string | undefined;
-const NODE_ENV: Env = process.env.NODE_ENV;
+const SVG = dynamic(() => import('react-inlinesvg'));
 
-const Image = React.forwardRef<any, ImageProps>(
-  (
-    {
-      src,
-      className,
-      media,
-      columns = 12,
-      priority = true,
-      itemProp,
-      testId,
-      disableInlineSVG,
-      nextImageOptimization,
-      q,
-      unoptimized,
-      width,
-      height,
-      layout,
-      ...imageProps
-    },
-    ref
-  ) => {
-    if (!src) return null;
-    const isSVG = src?.includes('.svg');
+const Image = React.forwardRef<any, ImageProps>((props, ref) => {
+  const {
+    src,
+    className,
+    media,
+    columns = 12,
+    priority = true,
+    itemProp,
+    testId,
+    disableInlineSVG,
+    nextImageOptimization,
+    q,
+    unoptimized,
+    width,
+    height,
+    alt,
+    ...imageProps
+  } = props;
+  if (!src) return null;
+  const imgPreload = React.useMemo(
+    () => (
+      <Head>
+        <link
+          rel="preload"
+          href={src}
+          // @ts-ignore
+          imagesrcset={getImgSrcTag({ src, numColumns: columns, q, unoptimized })?.srcSet}
+          as="image"
+          media={media}
+        />
+      </Head>
+    ),
+    [src, columns, q, unoptimized]
+  );
+  const imgContent = React.useMemo(() => {
+    const isSVG = src?.endsWith('.svg');
 
     let content;
     if (isSVG && !disableInlineSVG) {
-      content = <SVG innerRef={ref} src={src} data-testid={testId} className={className} {...(imageProps as any)} />;
+      content = (
+        <>
+          <SVG innerRef={ref} src={src} data-testid={testId} className={className} {...(imageProps as any)} />
+          {imgPreload}
+        </>
+      );
     } else if (!isSVG && nextImageOptimization) {
-      return (
+      content = (
         <NextImage
           src={src}
           // TODO: NextImage doesn't support ref
@@ -51,51 +67,34 @@ const Image = React.forwardRef<any, ImageProps>(
           loading={priority ? 'eager' : 'lazy'}
           height={height}
           width={width}
-          layout={layout}
+          sizes={imageProps.sizes}
+          alt={alt}
           {...imageProps}
         />
       );
     } else {
       content = (
-        <img
-          {...getImgSrcTag({ src, numColumns: columns, q, unoptimized })}
-          ref={ref}
-          data-testid={testId}
-          className={className}
-          itemProp={itemProp}
-          loading={priority ? 'eager' : 'lazy'}
-          {...imageProps}
-        />
+        <>
+          <img
+            {...getImgSrcTag({ src, numColumns: columns, q, unoptimized })}
+            ref={ref}
+            data-testid={testId}
+            className={className}
+            itemProp={itemProp}
+            loading={priority ? 'eager' : 'lazy'}
+            height={height}
+            width={width}
+            alt={alt}
+            {...imageProps}
+          />
+          {imgPreload}
+        </>
       );
     }
-    try {
-      return (
-        <ErrorBoundary>
-          {NODE_ENV === 'test' || priority ? (
-            <>
-              {content}
-              {isSVG ? (
-                <Head>
-                  <link
-                    rel="preload"
-                    href={src}
-                    // @ts-ignore
-                    imagesrcset={getImgSrcTag({ src, numColumns: columns, q, unoptimized })?.srcSet}
-                    as="image"
-                    media={media}
-                  />
-                </Head>
-              ) : null}
-            </>
-          ) : (
-            <LazyLoad offset={typeof window === 'undefined' ? 1000 : window.innerHeight}>{content}</LazyLoad>
-          )}
-        </ErrorBoundary>
-      );
-    } catch (err) {
-      return null;
-    }
-  }
-);
+    return content;
+  }, [props]);
 
-export default React.memo(Image);
+  return <ErrorBoundary>{imgContent}</ErrorBoundary>;
+});
+
+export default Image;