+++ packages/component-library/src/components/NavigationItem/NavigationItem.tsx
@@ -7,16 +7,16 @@ import { useMediaQuery } from '@mui/material';
 import { useTheme } from '@mui/system';
 
 import ErrorBoundary from '../ErrorBoundary';
-import Link from '../Link';
+import { LinkProps } from '../Link';
 import ContentModule from '../ContentModule';
-import sidekick from '../../utils/sidekick';
+import sidekick from '@last-rev/contentful-sidekick-util';
 import { NavigationItemProps } from './NavigationItem.types';
 
 export const NavigationItem = ({ subNavigation, sidekickLookup, onRequestClose, ...props }: NavigationItemProps) => {
   const [open, setOpen] = React.useState<boolean>(false);
   const theme = useTheme();
   const menuBreakpoint = theme?.components?.Header?.mobileMenuBreakpoint ?? 'sm';
-  const isMobile = useMediaQuery(theme.breakpoints.down(menuBreakpoint));
+  const isMobile = useMediaQuery(theme.breakpoints.down(menuBreakpoint), { defaultMatches: true });
   const handleClick = (evt: any) => {
     if (isMobile && subNavigation?.length) {
       evt.preventDefault();
@@ -36,16 +36,25 @@ export const NavigationItem = ({ subNavigation, sidekickLookup, onRequestClose, 
   return (
     <ErrorBoundary>
       <Root sx={{ position: 'relative' }} open={open} data-testid="NavigationItem" menuBreakpoint={menuBreakpoint}>
-        <Link {...props} {...sidekick(sidekickLookup)} onClick={handleClick} />
+        <NavigationItemLink
+          {...(props as LinkProps)}
+          {...sidekick(sidekickLookup)}
+          onClick={handleClick}
+          __typename="Link"
+        />
         {subNavigation?.length ? (
-          <MenuRoot menuBreakpoint={menuBreakpoint}>
+          <MenuRoot menuBreakpoint={menuBreakpoint} component={'ul'}>
             {subNavigation?.map((item) => (
               <MenuItem key={item.id}>
                 <ContentModule
                   {...item}
                   variant={'link'}
                   onClick={handleSubnavClick}
-                  onRequestClose={handleSubnavClick}
+                  {...(item?.__typename == 'NavigationItem'
+                    ? {
+                        onRequestClose: handleSubnavClick
+                      }
+                    : {})}
                 />
               </MenuItem>
             ))}
@@ -60,11 +69,13 @@ const visibleStyles = (open: boolean) => `
   max-height: ${open ? 300 : 0}px;
   box-shadow: ${open ? 'inset 0 0 16px -8px rgb(0 0 0 / 30%)' : 'inset 0 0 0 0 rgb(0 0 0 / 0%)'};
 `;
+const shouldForwardProp = (prop: string) =>
+  prop !== 'variant' && prop !== 'onRequestClose' && prop !== 'menuBreakpoint';
 
 const Root = styled(Box, {
   name: 'NavigationItem',
   slot: 'Root',
-  shouldForwardProp: (prop) => prop !== 'variant',
+  shouldForwardProp,
   overridesResolver: (_, styles) => [styles.root]
 })<{ variant?: string; open: boolean; menuBreakpoint: 'xs' | 'sm' | 'md' | 'lg' | 'xl' }>`
   ${({ open, theme, menuBreakpoint }) => `
@@ -87,12 +98,19 @@ const Root = styled(Box, {
   `}
 `;
 
+const NavigationItemLink = styled(ContentModule, {
+  name: 'NavigationItem',
+  slot: 'Link',
+  shouldForwardProp,
+  overridesResolver: (_, styles) => [styles.link]
+})<LinkProps>``;
+
 const MenuRoot = styled(Paper, {
   name: 'NavigationItem',
   slot: 'MenuRoot',
-  shouldForwardProp: (prop) => prop !== 'variant',
+  shouldForwardProp,
   overridesResolver: (_, styles) => [styles.menuRoot]
-})<{ variant?: string; menuBreakpoint: 'xs' | 'sm' | 'md' | 'lg' | 'xl' }>`
+})<{ variant?: string; menuBreakpoint: 'xs' | 'sm' | 'md' | 'lg' | 'xl'; component?: string }>`
   ${({ theme, menuBreakpoint }) => `
     display: flex;
     flex-direction: column;
@@ -124,7 +142,7 @@ const MenuRoot = styled(Paper, {
       .MuiMenuItem-root{
         width: 100%;
         display: block;
-        padding: 0;
+        // padding: 0;
         > div {
           width: 100%;
         }