+++ packages/component-library/src/components/SafeHTML/SafeHTML.types.ts
@@ -1,0 +1,3 @@
+export interface SafeHTMLProps {
+  children: string;
+}