+++ packages/component-library/src/components/SafeHTML/index.ts
@@ -1,0 +1,2 @@
+export { default } from './SafeHTML';
+export type { SafeHTMLProps } from './SafeHTML.types';