+++ packages/contentful-redis-loader/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @last-rev/contentful-redis-loader
 
+## 0.3.3
+
+### Patch Changes
+
+- 3ba98cd: Bump testing-library version
+- Updated dependencies [3ba98cd]
+  - @last-rev/timer@0.1.3
+
 ## 0.3.2
 
 ### Patch Changes