+++ packages/contentful-sidekick-util/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @last-rev/contentful-sidekick-util
 
+## 0.1.1
+
+### Patch Changes
+
+- 3ba98cd: Bump testing-library version
+
 ## 0.1.0
 
 ### Minor Changes