+++ packages/timer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @last-rev/timer
 
+## 0.1.3
+
+### Patch Changes
+
+- 3ba98cd: Bump testing-library version
+
 ## 0.1.2
 
 ### Patch Changes